refactor: changes paths of type.ts
