+++ src/utils/delFromLocalStorage.ts
@@ -1,4 +1,4 @@
-import { ICartLot } from '../styles/types';
+import { ICartLot } from '../types/types';
 
 export default function delFromLocalStorage(id: number) {
     const storageArray: ICartLot[] = JSON.parse(localStorage.cart);
