+++ src/utils/getTotallyCartCount.ts
@@ -1,4 +1,4 @@
-import { ICartLot } from '../styles/types';
+import { ICartLot } from '../types/types';
 
 export default function getTotallyCartCount() {
     if (localStorage.cart && localStorage.cart != null) {
